refactor: update public route handling and restructure user profile types
- Added '/profile/' to the list of public routes in middleware for improved access control.
- Re-exported UserProfileData type from a new profile.ts file for better organization and backward compatibility.
- Cleaned up imports in profile actions to reference the new UserProfileData location, enhancing code clarity.
