+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "idling.app",
-  "version": "0.36.3",
+  "version": "0.36.4",
   "private": true,
   "engines": {
     "node": ">=20.x"
