+++ src/app/components/user-profile/UserProfile.tsx
@@ -1,24 +1,12 @@
 'use client';
 
 import { useEffect, useState } from 'react';
+import { UserProfileData } from '../../../lib/types/profile';
 import { Avatar } from '../avatar/Avatar';
 import './UserProfile.css';
 
-export interface UserProfileData {
-  id: string;
-  username?: string;
-  name?: string;
-  bio?: string;
-  location?: string;
-  email?: string;
-  image?: string;
-  created_at?: string;
-  profile_public?: boolean;
-  total_submissions?: number;
-  posts_count?: number;
-  replies_count?: number;
-  last_activity?: Date | null;
-}
+// Re-export for backward compatibility
+export type { UserProfileData };
 
 interface UserProfileProps {
   user: UserProfileData;
